requirements.txt
dj_database_url
django
django-debug-toolbar
psycopg2
